Initial Commit
This is the first version of the algorithm which generates CWLE's for the Forward-Forward algorithm applied to CNNs. Development is currently in the works for a GUI application that allows one to more easily generate CWLE's and create them by hand while also keeping an eye on their convergence process. More inspection must be done to ensure that this algorithm meets the necessary requirements.